+++ configerz/tests/test_json.py
@@ -1,6 +1,6 @@
 import json
 
-from configerz import JSON_Controller, Configuration
+from configerz import JSON_Configuration
 
 from .core import BaseConfigTest, configuration_file_path,\
     default_configuration_dict, default_cfg_json_path
@@ -12,9 +12,7 @@ class Test_Create_DefDict(BaseConfigTest):
     and get default config values from `default_configuration_dict`
     """
     def setup(self):
-        self.config = Configuration()
-        self.controller = JSON_Controller(
-            self.config,
+        self.config = JSON_Configuration(
             configuration_file_path,
             default_configuration_dict
         )
@@ -29,9 +27,7 @@ class Test_Create_DefPath(BaseConfigTest):
         with open(default_cfg_json_path, 'w') as f:
             json.dump(default_configuration_dict, f)
 
-        self.config = Configuration()
-        self.controller = JSON_Controller(
-            self.config,
+        self.config = JSON_Configuration(
             configuration_file_path,
             default_cfg_json_path
         )